Revert "#9 refactor routes"
