+++ .gitignore
@@ -1,3 +1,5 @@
+notes.md
+
 node_modules
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
