+++ src/App.test.js
@@ -1,6 +1,6 @@
 // import { render, screen } from "@testing-library/react";
 // import App from "./App";
 
-test("renders learn react link", () => {
+test("renders", () => {
     expect(true).toBe(true);
 });
